Split http module into files
It was too long
