+++ frontend/src/services/api/user.js
@@ -1,5 +1,5 @@
 import store from '../../store';
-import { GET } from '../../store/modules/http';
+import { GET } from '../../store/modules/http/actions';
 
 export function getMyself() {
   return store.dispatch(`http/${GET}`, { url: 'users/me' });
