+++ frontend/src/store/modules/http/state.js
@@ -1,0 +1,8 @@
+export const TOKEN = 'TOKEN';
+
+export default {
+  // LEGACY SUPPORT
+  // After merge to new version remove localStorage
+  [TOKEN]: localStorage.getItem('jwt') || null,
+};
+
